separate DependencyConstraint and Dependency declarations
The interfaces for declaring these two different things were
coupled in the initial implementation - to reuse all functionality
based on the Dependency interface directly. This interface is used
internally to pass dependency declarations through the resolution
process. However, this is mostly due to the "old" results API
(Configuration.resolvedConfiguration()).

These implementation specifics should not bleed into the API.

A dependency declaration defines a *requirement*:
  I require - some version - of module X

A dependency constraint defines a *constraint*:
  If I must use X, I can only work with versions matching the constraint

These are two different things conceptually and in the DSL. Only
internally they are (currently) represented by the same data structure.
Which might also change in the future.