Include dependency constraints in resolution result
The "old" results API has methods using the "Dependency"
interface. Although a result could now as well be based on a
DepencdencyConstraint.

This commit is a hack to test this implementation. We can not keep it.
We can probably accept that the "old" resolution result API can not be
used to access constraints. And let users use the "new"
'configuration.incoming.resolutionResult' API instead.

If we remove this commit, we need to adjust 'ResolveTestFixture' for
dependency constraints tests. The fixture heavily relies on (and tests)
the "old" results API.